[Wave] Refactor attention lit tests
This PR breaks the attention lit tests into
separate files and adds them in a new subdirectory.
